@@ -1,7 +1,7 @@
 import 'dart:html';
 
 /// Create anchor element with download attribute
-AnchorElement createAnchorElement(String href, String suggestedName) {
+AnchorElement createAnchorElement(String href, String? suggestedName) {
   final element = AnchorElement(href: href);
 
   if (suggestedName == null) {
@@ -27,7 +27,7 @@ Element ensureInitialized(String id) {
   if (target == null) {
     final Element targetElement = Element.tag('flt-x-file')..id = id;
 
-    querySelector('body').children.add(targetElement);
+    querySelector('body')!.children.add(targetElement);
     target = targetElement;
   }
   return target;
