@@ -23,6 +23,7 @@ import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.PatternItem;
 import com.google.android.gms.maps.model.RoundCap;
 import com.google.android.gms.maps.model.SquareCap;
+import com.google.android.gms.maps.model.Tile;
 import io.flutter.view.FlutterMain;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -78,7 +79,8 @@ class Convert {
       }
     } else {
       throw new IllegalArgumentException(
-          "fromBytes should have exactly one argument, the bytes. Got: " + data.size());
+          "fromBytes should have exactly one argument, interpretTileOverlayOptions the bytes. Got: "
+              + data.size());
     }
   }
 
@@ -197,6 +199,20 @@ class Convert {
     }
     final Map<String, Object> data = new HashMap<>(1);
     data.put("circleId", circleId);
+    return data;
+  }
+
+  static Map<String, Object> tileOverlayArgumentsToJson(
+      String tileOverlayId, int x, int y, int zoom) {
+
+    if (tileOverlayId == null) {
+      return null;
+    }
+    final Map<String, Object> data = new HashMap<>(4);
+    data.put("tileOverlayId", tileOverlayId);
+    data.put("x", x);
+    data.put("y", y);
+    data.put("zoom", zoom);
     return data;
   }
 
@@ -644,5 +660,40 @@ class Convert {
       default:
         throw new IllegalArgumentException("Cannot interpret " + o + " as Cap");
     }
+  }
+
+  static String interpretTileOverlayOptions(Map<String, ?> data, TileOverlaySink sink) {
+    final Object fadeIn = data.get("fadeIn");
+    if (fadeIn != null) {
+      sink.setFadeIn(toBoolean(fadeIn));
+    }
+    final Object transparency = data.get("transparency");
+    if (transparency != null) {
+      sink.setTransparency(toFloat(transparency));
+    }
+    final Object zIndex = data.get("zIndex");
+    if (zIndex != null) {
+      sink.setZIndex(toFloat(zIndex));
+    }
+    final Object visible = data.get("visible");
+    if (visible != null) {
+      sink.setVisible(toBoolean(visible));
+    }
+    final String tileOverlayId = (String) data.get("tileOverlayId");
+    if (tileOverlayId == null) {
+      throw new IllegalArgumentException("tileOverlayId was null");
+    } else {
+      return tileOverlayId;
+    }
+  }
+
+  static Tile interpretTile(Map<String, ?> data) {
+    int width = toInt(data.get("width"));
+    int height = toInt(data.get("height"));
+    byte[] dataArray = null;
+    if (data.get("data") != null) {
+      dataArray = (byte[]) data.get("data");
+    }
+    return new Tile(width, height, dataArray);
   }
 }
