@@ -42,7 +42,11 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         PolygonId,
         Polyline,
         PolylineId,
-        ScreenCoordinate;
+        ScreenCoordinate,
+        Tile,
+        TileOverlayId,
+        TileOverlay,
+        TileProvider;
 
 part 'src/controller.dart';
 part 'src/google_map.dart';
